+++ src/MBW.Utilities.Journal.Tests/JournaledStreamExamples.cs
@@ -14,7 +14,7 @@ public class JournaledStreamExamples
             using (FileStream fileStream = new FileStream(filePath, FileMode.OpenOrCreate, FileAccess.ReadWrite))
             {
                 // Wrap the file stream in a JournaledStream. This allows us to read and write to the stream without affecting the underlying file.
-                using (JournaledStream journalStream = new JournaledStream(fileStream, journalPath))
+                using (JournaledStream journalStream = JournaledStreamFactory.CreateWalJournal(fileStream, journalPath))
                 {
                     // Write data to the JournaledStream
                     // Note! Do not use the original FileStream directly, as this bypasses the Journal
@@ -34,7 +34,7 @@ public class JournaledStreamExamples
             {
                 // Even though we will not write to the file, we still need to use the JournaledStream. This ensures
                 // that if our write was partial, we will replay the journal again to ensure it has been fully written.
-                using (JournaledStream journalStream = new JournaledStream(fileStream, journalPath))
+                using (JournaledStream journalStream = JournaledStreamFactory.CreateWalJournal(fileStream, journalPath))
                 {
                     // Read the data back from the JournaledStream
                     byte[] readBuffer = new byte[1024];
@@ -43,7 +43,7 @@ public class JournaledStreamExamples
 
                     Assert.Equal("Hello, Journaled World!", readData);
                 }
-                
+
                 // Alternatively, we can ensure the journal has been applied, by using the utility
                 // After this utility has run, we know that any journal has either been applied fully or removed (if it wasn't committed by the original application)
                 JournaledUtilities.EnsureJournalCommitted(fileStream, journalPath);
