+++ src/MBW.Utilities.Journal/Structures/IStructWithMagic.cs
@@ -1,0 +1,13 @@
+using System.Numerics;
+
+namespace MBW.Utilities.Journal.Structures;
+
+internal interface IStructWithMagic<out TMagic> : IStructWithMagic where TMagic : INumber<TMagic>
+{
+    public TMagic Magic { get; }
+}
+
+internal interface IStructWithMagic
+{
+    public static abstract int StructSize { get; }
+}
