+++ src/MBW.Utilities.Journal/Structures/JournalFileConstants.cs
@@ -1,0 +1,9 @@
+﻿namespace MBW.Utilities.Journal.Structures;
+
+internal static class JournalFileConstants
+{
+    /// <summary>
+    /// "JRNLVER1"
+    /// </summary>
+    public const ulong HeaderMagic = 0x4A524E4C56455231;
+}
